* Moving imtes to `utils.soures`. * Moving some of the helper utils to `panoptes.utils.bayer`
